internal/acctest: Add acctest logic into separate package
